Pick out correct files from filesystem
This uses scandir, which uses a filter and sort cmp function.  The
filter function discards files from the directory which we're not
interested in, in this case, anything that isn't a file from memdiff.
The filter function provides a hazy match on what is and isn't a memdiff
file, but should be "good enough"(tm) for most purposes I currently care
about right at this particular instant in time.  It should be noted that
the filter function totally breaks all expectations and sets global
variables within it because side-effect free programming is a great
idea... for people who just don't enjoy life anymore and want their
programs to be reliable instead of interesting.

The sort comparison function is some of the worst code I've written in
recent memory.  It should work though.  It sorts the files by segment,
then by snapshot.  But really this code should have been written in
basically any other language.  Anything with high level string functions
would have been fine.  Hell, even a better string handling library in C
would have dealt with this better.

Basically I needed to do parsing in this code, so I took down the
shotgun and went to it.  So.  Sorry about that shotgun parser.  It
probably works though.

I wouldn't use it on untrusted filesystems.