Basic functionality
There's some rough spots in this code, but it mostly works and provides
most of what needs to be visualized.  There's definitely some
interesting errors that happens with byte boundaries, since we don't
record how many blocks a memdiff file actually is, it's impossible to
recover that information just looking at the file in the current format,
since file *must* contain full bytes and can't truncate bits mid-byte,
which means we can't really determine where the blocks end.

The right way to fix this is to change the memdiff format and give it a
count of the blocks at the beginning, but that's going to have to wait
until the next version.